Refactor permission handling

+++ backend/src/main/java/hu/bsstudio/raktr/comment/service/CommentService.java
@@ -2,6 +2,7 @@ package hu.bsstudio.raktr.comment.service;
 
 import hu.bsstudio.raktr.dal.repository.CommentRepository;
 import hu.bsstudio.raktr.exception.ObjectNotFoundException;
+import hu.bsstudio.raktr.security.SecurityService;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
@@ -14,9 +15,13 @@ public class CommentService {
 
     private final CommentRepository commentRepository;
 
+    private final SecurityService securityService;
+
     @Transactional
     public void deleteComment(Long commentId) {
         var comment = commentRepository.findById(commentId).orElseThrow(ObjectNotFoundException::new);
+
+        securityService.checkIsOwnerOrAdmin(comment.getCreatedBy().getUuid());
 
         commentRepository.delete(comment);
